Update stream.php
Same times not on all streams the client stream crash ( if the stream fils is 19 or change 19 to 20 this happen, as example parallel on stream get black pic and image again this make the client stream crash  ). This affect the $active_cons also and make the clients no respect the user limit. The comment on //++$try fix that behavior and client never crash, but if your channel go down or stopped by you, have to clear the activity log to get the client on line again. I have no idea how ti fix that is my first tach with php staff and if any one helps is very good :) thanks